添加文档发送、修改课程表

+++ app/src/main/java/com/bll/lnkteacher/ui/fragment/document/DocumentFragment.kt
@@ -3,21 +3,49 @@ package com.bll.lnkteacher.ui.fragment.document
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.recyclerview.widget.GridLayoutManager
+import com.bll.lnkteacher.DataBeanManager
 import com.bll.lnkteacher.FileAddress
 import com.bll.lnkteacher.MethodManager
 import com.bll.lnkteacher.R
 import com.bll.lnkteacher.base.BaseFragment
-import com.bll.lnkteacher.dialog.CommonDialog
+import com.bll.lnkteacher.dialog.ClassGroupSelectorDialog
+import com.bll.lnkteacher.dialog.LongClickManageDialog
+import com.bll.lnkteacher.mvp.model.ItemList
+import com.bll.lnkteacher.mvp.presenter.DocumentPresenter
+import com.bll.lnkteacher.mvp.presenter.QiniuPresenter
+import com.bll.lnkteacher.mvp.view.IContractView
 import com.bll.lnkteacher.ui.adapter.DocumentAdapter
 import com.bll.lnkteacher.utils.DP2PX
+import com.bll.lnkteacher.utils.FileUploadManager
 import com.bll.lnkteacher.utils.FileUtils
 import com.bll.lnkteacher.widget.SpaceGridItemDeco
 import kotlinx.android.synthetic.main.fragment_give_lessons.rv_list
-import java.io.File
 
-class DocumentFragment:BaseFragment() {
+class DocumentFragment:BaseFragment(),IContractView.IQiniuView,IContractView.IDocumentView {
+    private var qiniuPresenter=QiniuPresenter(this)
+    private var documentPresenter=DocumentPresenter(this,1)
     private var mAdapter: DocumentAdapter? = null
     private var path=""
+    private var classIds= mutableListOf<Int>()
+    private var position=0
+
+    override fun onToken(token: String) {
+        val file= mAdapter?.data?.get(position)
+        FileUploadManager(token).apply {
+            upload(file?.path!!)
+            setCallBack{
+                val map=HashMap<String,Any>()
+                map["title"]=FileUtils.getUrlName(file.path)
+                map["url"]=it
+                map["classIds"]=classIds
+                documentPresenter.sendDocument(map)
+            }
+        }
+    }
+
+    override fun onSendSuccess() {
+        showToast(1,"发送成功")
+    }
 
     override fun getLayoutId(): Int {
         return R.layout.fragment_list_content
@@ -52,23 +80,46 @@ class DocumentFragment:BaseFragment() {
                 MethodManager.gotoDocument(requireActivity(), file)
             }
             setOnItemLongClickListener { adapter, view, position ->
-                val file = data[position]
-                onLongClick(file)
+                this@DocumentFragment.position=position
+                onLongClick()
                 true
             }
         }
     }
 
-    private fun onLongClick(file: File) {
-        CommonDialog(requireActivity(),1).setContent(R.string.toast_is_delete_tips).builder().setDialogClickListener(object : CommonDialog.OnDialogClickListener {
-            override fun ok() {
-                FileUtils.deleteFile(file)
-                val drawPath = file.parent + "/${FileUtils.getUrlName(file.path)}draw/"
-                FileUtils.delete(drawPath)
-                MethodManager.notifyFileScan(requireActivity(),file.absolutePath)
-                mAdapter?.data?.indexOf(file)?.let { mAdapter?.remove(it) }
-            }
+    private fun onLongClick() {
+        val file= mAdapter?.data?.get(position)!!
+        val beans = mutableListOf<ItemList>()
+        beans.add(ItemList().apply {
+            name = "删除"
+            resId = R.mipmap.icon_setting_delete
         })
+        beans.add(ItemList().apply {
+            name = "发送"
+            resId = R.mipmap.icon_setting_upload
+        })
+        LongClickManageDialog(requireActivity(),1, file.name, beans).builder()
+            .setOnDialogClickListener { position->
+                when(position){
+                    0->{
+                        FileUtils.deleteFile(file)
+                        val drawPath = file.parent + "/${FileUtils.getUrlName(file.path)}draw/"
+                        FileUtils.delete(drawPath)
+                        MethodManager.notifyFileScan(requireActivity(),file.absolutePath)
+                        mAdapter?.data?.indexOf(file)?.let { mAdapter?.remove(it) }
+                    }
+                    1->{
+                        if (!FileUtils.getUrlFormat(file.path).equals(".pdf")){
+                            showToast(1,"目前只支持发送pdf文件给学生")
+                            return@setOnDialogClickListener
+                        }
+                        ClassGroupSelectorDialog(requireActivity(),1,DataBeanManager.classGroups).builder().setOnDialogSelectListener{
+                            classIds= it.toMutableList()
+                            qiniuPresenter.getToken()
+                        }
+                    }
+                }
+            }
     }
 
     fun onTypeSelectorEvent(typeStr:String) {
@@ -87,4 +138,5 @@ class DocumentFragment:BaseFragment() {
     override fun onRefreshData() {
         fetchData()
     }
+
 }
